Sentinel L2A Imagery Download System
You can add an ESRI shape file, and run python ranger-app.py. The system will download
the most recent Sentinel L2A imagery tiles that includes the extent area of the polygons
in the provided shapefile. NDVI and RGB are calculated for the downloaded tile and
polygons are clipped and saved as individual rasters in a folder. They can also be
uploaded to Google Storage if a key file is provided.